证书相关方法挪到backend中, 保持root干净
